ChatBot/requirements.txt
openai
python-dotenv
psycopg2
fastapi
ollama